- revert library version bumps - fix: Convert id() Integers to Longs to avoid nfes

+++ src/main/java/com/spotify/github/v3/User.java
@@ -41,7 +41,7 @@ public interface User {
 
   /** ID. */
   @Nullable
-  Integer id();
+  Long id();
 
   /** Avatar URL. */
   @Nullable
