+++ src/main/java/com/spotify/github/v3/repos/RepositoryBase.java
@@ -40,7 +40,7 @@ public interface RepositoryBase extends UpdateTracking {
 
   /** ID */
   @Nullable
-  Integer id();
+  Long id();
 
   /** Name */
   @Nullable
